you can now exclude certain files like, ns_args
However, I can't figure out a way to directly exclude it in the unzip
package, hence, it just renames the original to "<file>.excluded" when
the extraction is done it then renames it back to it's original aka
"<file>", overwriting what was extracted, which essentially excludes
some files.

If there exists an unzip library/package that has options for excluding
files we should move to that, but until something as such is found the
current way is how we'll do it.