attempt at making localization
This may or may not be how we actually do localization in the future,
however for now this seems doable. I will obviously need to look at how
we detect the language, as I think instead of relying on names like
"en-US" just have "en", so we don't have to symlink various editions of
English to the same file. But for now this is a draft, and the important
part of this is rather how the underlying localization works.