buttons now disable when updating, and app logs
The buttons in the GUI disable whilst you're updating Northstar and
potentionally doing other things in the future, I also added a way to
log things in the app, albeit it just prints it in the "Welcome to
Viper!" part of the app, which is just fine.

I also added all the needed language strings for the GUI logs and
removed "gui.missinggamepath" as we use it for both the CLI and GUI even
tho "general.missinggamepath" exists, and so we now use the general one
for both, as the messages are the same.