--updatevp, and option to disable auto updates
If you want you can set "autoupdates" to false in your config, no GUI
tools to do this yet. For the CLI auto updates is off by default and
you'll have to use --updatevp.

I also removed the snap package, tho whether this stays as a change is
still to be discussed.

And with the new option I updated the help menu, the man page and
everything along else that needs it.