chore: Auto updating (#16)
* [chore] adding electron-updater dependency

* [feat] adding auto-updating mechanism

* [chore] setting package version to v0.8.0

* [fix] restoring original repo URL

* [docs] adding some documentation about publishing new releases

* [chore] adding publish:windows command

* [docs] updating publish instructions with new publish command

* [chore] adding publish:linux command

* [docs] updating publish instructions

* --updatevp, and option to disable auto updates

If you want you can set "autoupdates" to false in your config, no GUI
tools to do this yet. For the CLI auto updates is off by default and
you'll have to use --updatevp.

I also removed the snap package, tho whether this stays as a change is
still to be discussed.

And with the new option I updated the help menu, the man page and
everything along else that needs it.

* removed "soon" parts of README for auto-updates

* [feat] adding French translation for cli.help.updatevp key

* confirmation for restarting the app

Now instead of automatically updating and restarting the app, which may
be slightly confusing to some users, (the app opens then closes and then
opens), it now asks whether you want to restart and open the new
version. If you say no, instead next time you launch it, it'll be on the
new version.

If you want to completely disable updates you can disable it in the
viper.json file...

* [feat] adding French translation for gui.update.available key

* added configuration instruction in README

Co-authored-by: 0neGal <mail@0negal.com>