installing mods from folders got a lil easier
If a mod decided to put their actual mod files inside a folder inside a
folder it'll search sub folders to try and find them and usually it's
able to install it.

Meaning, if instead of the mod being structured as such:

mod/
    mod.json
    other files ...

It is structured as such:

folder/
       mod/
           mod.json
           other files ...

This will be very useful for zip files... As some people tend to archive
a folder and not make an archive with the files in it. Leading to the
above structure...