fixed/changed the README drastically
I somehow completely forgot to remove the "coming soon" parts for
auto-updating and official releases and download pages, that's now
fixed. And since we have that I also moved and changed some text in the
"How to run" section renaming it to "Development"

Besides that I added Viper's logo along with a link to the projects page
for a nice overview and releases page, even tho it is in fact in the
sidebar, you may miss it somehow.

Given I've removed the application preview we no longer have to keep
updating it everytime some change, changes how the app looks on startup.
Overall much needed changes to the README.