added utils.getSysLang()
This function spits out an IPC message on the "syslang" channel, with
the navigator.language, hence the system language. You're then free to
listen for that event and do something with it.

I also moved everything inside the app.on("ready") event which allows us
to call this from the CLI as well. Unfortunately that may come with tiny
slowdowns and possibly Chromium logs. Not much we can do about that.
However keep in mind these logs don't get passed through stdin, so they
won't affect automation scripts.