general improvements to the UI
Launch buttons now have a nice gradient, a shadow and a better hover
animation. The Titanfall logos are now also centered no matter the
window size, so if in the future we decide to allow resizing the window
that won't be a problem, or if you force the window to do it through
some trickery, it'll also still show correctly.

I also changed the default font weight to 500, some pieces of text
looked very thin.