feat: New UI (#26)
* [chore] removing setsize mechanism

* [feat] adding new html skeleton

* [feat] adding games icon buttons

* [feat] all icon buttons have same size

* [fix] games container width varies between 95 and 120px

* [feat] displaying bg image

* [feat] each game has its own background image

* [feat] adding content skeleton

* [feat] window has no frame anymore

* [feat] buttons show related content only

* [feat] displaying viper release notes

* [feat] game icons are vertically centered

* [feat] background is a bit grayed out

* [feat] adding some items to games menus

* [feat] displaying games logos

* [feat] adding content menu active item indicator

* [fix] centering active item indicator

* [feat] not displaying viper credits by default

* [feat] viper categories can be browsed through via menu

* [feat] adjusting hovering style of menu items

* [feat] adding play button styles

* [feat] adding missing vp sections

* [feat] adding link styles

* [fix] scaling down northstar logo

* [feat] adding ns sections

* [feat] fetching northstar release notes

* [fix] release notes containers are scrollable

* [feat] removing space between menu and some sections

* [feat] loading vp+ns versions into interface

* [chore] pushing version to 0.10.0

* [feat] play buttons launch games

* [refactor] renaming vp releases fetching method

* [feat] displaying an option to update northstar

* [feat] message is displayed if ns is up-to-date

* [feat] displaying ns download progress

* [refactor] removing ns-updating and ns-updated events

* [feat] adding UI translations

* [fix] correctly displaying vp release notes

* [refactor] removing useless code

* [feat] adding Imply to credits for viper logo

* [feat] loading titanfall version from gameversion.txt file

* [feat] centering viper main section

* [feat] adding custom scrollbars to release notes sections

* [feat] disabling images drag

* [feat] all external links are opened in default browser

* [feat] formatting release notes with markdown converter

* [feat] translating UI items

* fixed backgrounds not being same size

This would cause weird transitions.

* new.css is now main.css, fixed colors and fonts

* page indicator is now prettier

We no longer use a character to render a circle, and it's no longer a
circle, instead it's a line, we also now have a nice transition on it,
both for fading in and out.

* removed simplebar in favor of ::-webkit-scrollbar

Almost everything we did with simplerbar could be done with native
WebKit scollbar CSS, hence why we switched to it.

* fixed uneven nav button margin

* fixed formatting

Single quotes to double quotes, spaces to tabs, etc etc... I also
removed simplerbar properly...

* removed unneeded consts

Elements set with ID's can by default be referred to with their ID, aka:

    const element = document.getElementById("element");

Is useless as "element" will automatically refer to that element if
theres no variable that overwrites it.

* use release name over tag_name

"Release vX.X.X" frankly looks better than "vX.X.X"

* playBtn's now look nicer

Slight transition on hover and click.

* added pointer cursors to nav buttons and playBtn's

Along with transitions to the nav buttons, because they look very good.

* [fix] viper main section appears correctly on viper logo clicked

* different opacity for inactive pages

* removed "Roboto Mono", added normal "Roboto"

* removed old HTML

* fixed formatting in launcher.js

* redesigned contentContainer's, and a lot more

I also swapped around the Vanilla logo and Northstar logo to have the
same size and the "Titanfall 2" text is in the same place on both
images, allowing seamless transition when changing pages.

I also completely redesigned displayContent(), it's now page() and
doesn't revolve around a switch case. I'll likely do a similar thing
with showVpSection() and showNsSection(), uniting them into one function
that doesn't use switch cases.

* added close button

* fix game icons not being found

* removed excess icons from assets folder

* [feat] adding a br to french releases key

* [feat] centering menu items

* [feat] implementing requests cache for NS release notes

* [feat] implementing cache for VP release notes

* changed ttf2 to tf2

Along with that, I also renamed the northstar2.jpeg bg to just
northstar.jpg, as all other background are .jpg

* [fix] adjusting containers

* [fix] adding some space between menu and content

Since we have lots of space to use now, let's use it!
This extra-space does not apply to release notes sections, which
already takes all available screen space.

* [feat] adding code style

* [fix] window is draggable thanks to a drag bar on its top

* [fix] drag bar height

* [fix] all window is draggable

* [fix] not disabling game buttons on ns update

* [fix] disabled buttons cannot be clicked

* [fix] displaying ns update downloading state on UI

* [fix] typos

* general improvements to the UI

Launch buttons now have a nice gradient, a shadow and a better hover
animation. The Titanfall logos are now also centered no matter the
window size, so if in the future we decide to allow resizing the window
that won't be a problem, or if you force the window to do it through
some trickery, it'll also still show correctly.

I also changed the default font weight to 500, some pieces of text
looked very thin.

* unified information and credits section

* mods section should more or less look good now

Still not complete done, I still wanna implement a drag and drop
install, along with selecting whether you're installing a zip or folder.
And also a warning when toggling or removing required mods...

But generally this is the design I will go with.

* added warnings for toggling/removing core mods

* [feat] adding missing translations

* [fix] typo in gui.mods.disabledtag french translation

* re-added download progress

* added section transitions

* switched to npm module for markdown parsing

* correcting some formatting, nothing special

* active page button is now more obvious

Co-authored-by: 0neGal <mail@0negal.com>