added loading, close button, element fadein etc
To show the loading simply do Browser.loading(), when a BrowserEl() is
made it auto removes it. The loading and BrowserEl's will fade into
view instead of popping in instantly.

And then now there's a tiny close button, as it may not be obvious you
can click the dark overlay or click Escape.