no alert when manually changing the game path
When changing the game path by clicking the button you shouldn't be told
it can't find the game and you've to select one manually, as you know
that already. More importantly, if it could be found automatically it'll
just not do anything.

With this change I also changed the "gui.setpath" string to be more
logical, and to make it clear what it does.