various changes, aesthetic and functionality
This should fix the window drag not working, however I cannot test that.
As my WM doesn't support it, a Windows contributor will have to test it.

I also added an info button that opens the mod's page in your web
browser, this button is of course localized, along with that I localized
the placeholder in the search input.

There's also a few general aesthetic changes, like adding a tiny border
around #modsdiv and #browser.