fixed drag control
I think? I don't know how it was originally, currently now you can drag
on the top most of the window and in the left most, unfortunately not
inside the various pages.

It would probably be possible to do that, but I've not made it do as
such, and if it needs to be changed it won't be me doing it.