added support for more obscure mod structure
If a developer for some reason decides to put their mod inside mods/
instead of putting it inside a folder in mods/, it used to fail the
install, now that's fixed.

However since there's no folder to directly copy the name from we take
it from the manifest.json provided by Thunderstore.

It of course also has the proper status in the browser UI
(update/reinstall etc)