Thunderstore mods are only loaded when requested
Before this commit, the Thunderstore request would be sent on
application startup, now instead, it only does it when someone actually
clicks the "Find Mods" button, which should intern decrease load on
Thunderstore's end. Not that it mattered in the grand scheme of things,
it was just sort of unneeded.