check to make sure gamepath exists
When Viper starts up it'll check to make sure the gamepath still exists,
and throws errors if not, it also redirects you to the first page (the
one where you can set the gamepath), and gives you an informative error.

This could happen because the user unmounted the drive the gamepath is
on, or it could happen if the user moved their game location.