simplified get() and made list() used get()
Instead of repeating code and everything, get() now only returns a
boolean, based on the arg inputted into get(mod), it returns true if the
mod is enabled and false if disabled.

If the mod isn't found inside the enabledmods.json it is assumed that
it's enabled, as Northstar doesn't update this unless you actually
disable the mod first, and the mod won't be found in there.

Then I used get() to simplify list() a whole lot, and adding that
functionality of using the manifest for a version number if one couldn't
be found in the mod.json.