already installed dependencies are now skipped
If you already have all the dependencies or some dependencies of a
package those will be skipped, if there is no dependencies missing it'll
just install, and otherwise it'll show the missing and ask whether you
want to install them.

Meaning if a package has two dependencies and you've one of them only
the one you don't have will show up.