a lot of stuff, but mainly a "Load more..." button
In the past when more than 50 packages was loaded it'd simply display a
"Maximum packages loaded" message, and it wasn't actually 50, as it just
took the first 50 packages loaded, whether or not they were filtered
out, so now that's fixed and a "Load more..." button has been added.

With that comes some changes to the lang strings for
"gui.browser.endoflist" and a new one "gui.browser.loadmore".

I also fixed the filtered packages issue also occurring in searches.

This will overall allow a user to more easily find the maximum amount of
packages without loading literally over a hundred packages at once, not
good for performance...