removed cookie and manager alert in webview
When opening the preview for a mod it may show a cookie alert, this is
now auto-hidden, along with an alert recommending users to use the
Thunderstore Mod Manager, or an alternative, since Viper is an
alternative, this is only an alert that can confuse users. And it is
therefore removed.