added more user friendly error message
When JavaScript errors occur outside of the renderer, they'll no longer
display a big and confusingly detailed error stack, now they'll simply
be shown a toast about the fact that an error happened.

The user can then click this to get more details, but still without it
being as invasive and obtuse as before.