different colored switches
When a red switch is placed besides a red button that says "Remove", it
makes it seem far more destructive to click that switch, even if it's
not, and a blue color doesn't fit in very well either (perhaps we need
more colors). And so because of all that the modlist switches are now
orange, and we can easily add more colors.

The switch colors overall are also far more vibrant now, as it looks far
better that way.