win.alert() now uses Promise, and added .confirm()
win.alert() now has a Promise return value, which'll allow you to wait
until an alert has been closed before continuing code execution.

win.confirm() was also added it's the same as win.alert() except it runs
confirm() in the renderer instead of alert(), and has a return boolean
in the Promise resolve callback. This boolean being whether the user
confirmed the action or not.