added hover tooltips
They're pretty simple to use, and automatically make themselves fit on
screen, on top of automatically disappearing and appearing of course.

I've used them a few places now, and not sure where else they can be
used, there are of course localization strings that need, well,
localization, and I'll create a PR for this on a later date.