fixed settings file refusing to save if invalid
This is quite counter intuitive, as, if it's invalid, how do we
re-validate it, if we can't even save it again.

This existed here because the idea was for the re-validation to occur
somewhere else, and to make sure incorrectly formatted data wasn't being
given to this function. Now the function simply resets the config, a
restart may be required to add missing settings, in case the parsed
`conf` doesn't have all keys.