fixed gamepath buttons being disabled
If the gamepath is lost or similar then it'll disable many buttons,
notably the install/launch buttons and other similar buttons, however
for obvious reasons we shouldn't be stopping the user from changing
their gamepath in this scenario. This is still useful when installing a
mod or updating NS, so it's just this scenario.