fixed "Force quit game" button getting disabled
Due to this button (specifically the "button"/link below the "Launch"
button) serving multiple functions, it incorrectly gets disabled when
the game is running. Now its enabled!

How did I miss this? I assume it crept up after implementing the force
quit button, and not during, but maybe not, was I stupid? Mayhaps.