requests.js ignores cache time when no internet
This means if you dont have any internet, but there's a cached requests,
it'll use that, even if it was cached a very long time ago. This just
attempts to eliminate errors.

This can still be turned off for things where you dont want this to
happen, notably the masterserver status.