fixed setting cache folder too late
We set the Electron path `userData` to the system's cache folder, as
this is the same path Electron stores cache files in, and its normally
set to the system config folder, obviously not ideal.

However, when changing, we do as such quite late in the startup process
leading to some things (notably the variable used to store where we
download the Northstar archive to), using the older value.

Now its one of the first things we do on startup, fixing that.